+++ packages/terra-application-navigation/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * `isHeaderDisabled` to hide header from popup.
+
 ## 1.85.0 - (May 14, 2024)
 
 * Changed