+++ packages/terra-framework-docs/CHANGELOG.md
@@ -10,6 +10,9 @@
 ## 1.87.0 - (May 14, 2024)
 
 * Changed
+  * Enabled `close` button in examples.
+  * Modified Button in `PopupStandard` example.
+  * Removed `ariaLabel` from `PopupStandard` example.
   * Updated `terra-slide-panel` example to use `terra-buttons`.
 
 ## 1.86.0 - (May 8, 2024)