+++ packages/terra-menu/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added `isMenu` prop to the `Popup` component props.
+
 ## 6.97.0 - (May 14, 2024)
 
 * Changed