+++ packages/terra-popup/CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-* Changed
-  * Changed `targetRef` to an optional prop.
-
 ## Unreleased
+
+* Changed
+  * Enabled `close` button in popup by default.
 
 ## 6.83.0 - (May 14, 2024)